Fix up tests after resolving merge conflicts
