+++ src/tests.rs
@@ -1,6 +1,5 @@
 mod test_bits;
 mod test_cell_path;
-mod test_cell_paths;
 mod test_conditionals;
 mod test_config_path;
 mod test_converters;
