+++ src/tests/test_cell_path.rs
@@ -1,6 +1,11 @@
 use crate::tests::{fail_test, run_test, TestResult};
 
-// Tests for $nothing / null / Value::Nothing
+// tests for $nothing / null / Value::Nothing
+#[test]
+fn nothing_passes_optional() -> TestResult {
+    run_test("$nothing?.foo | to nuon", "null")
+}
+
 #[test]
 fn nothing_fails_string() -> TestResult {
     fail_test("$nothing.foo", "doesn't support cell paths")
@@ -28,6 +33,11 @@ fn record_int_failure() -> TestResult {
 }
 
 #[test]
+fn record_single_field_optional() -> TestResult {
+    run_test("{foo: 'bar'}?.foobar  | to nuon", "null")
+}
+
+#[test]
 fn nested_record_field_success() -> TestResult {
     run_test("{foo: {bar: 'baz'} }.foo.bar == 'baz'", "true")
 }
@@ -35,6 +45,11 @@ fn nested_record_field_success() -> TestResult {
 #[test]
 fn nested_record_field_failure() -> TestResult {
     fail_test("{foo: {bar: 'baz'} }.foo.asdf", "")
+}
+
+#[test]
+fn nested_record_field_optional() -> TestResult {
+    run_test("{foo: {bar: 'baz'} }.foo?.asdf  | to nuon", "null")
 }
 
 #[test]
@@ -52,7 +67,15 @@ fn record_with_nested_list_column_failure() -> TestResult {
     fail_test("{foo: [{bar: 'baz'}]}.foo.0.asdf", "")
 }
 
-// Tests for lists
+#[test]
+fn deeply_nested_optional_cell_path() -> TestResult {
+    run_test(
+        "{foo: [{bar: 'baz'}]}.foo?.3?.bar?.asdfdafg?.234?.foobar  | to nuon",
+        "null",
+    )
+}
+
+// tests for lists
 #[test]
 fn list_single_field_success() -> TestResult {
     run_test("[{foo: 'bar'}].foo.0 == 'bar'", "true")?;
@@ -65,17 +88,35 @@ fn list_single_field_failure() -> TestResult {
     fail_test("[{foo: 'bar'}].asdf", "")
 }
 
-// Test the scenario where the requested column is not present in all rows
 #[test]
-fn jagged_list_access_succeeds() -> TestResult {
-    run_test("[{foo: 'bar'}, {}].foo == ['bar', $nothing]", "true")?;
-    run_test("[{}, {foo: 'bar'}].foo == [$nothing, 'bar']", "true")
+fn list_single_field_optional() -> TestResult {
+    run_test("[{foo: 'bar'}]?.asdf.0  | to nuon", "null")?;
+    run_test("[{foo: 'bar'}].0?.asdf  | to nuon", "null")
+}
+
+// test the scenario where some list records are missing columns
+#[test]
+fn jagged_list_access_fails() -> TestResult {
+    fail_test("[{foo: 'bar'}, {}].foo", "")?;
+    fail_test("[{}, {foo: 'bar'}].foo", "")
+}
+
+// test the scenario where some list records are missing columns
+#[test]
+fn jagged_list_optional_access_succeeds() -> TestResult {
+    run_test("[{foo: 'bar'}, {}]?.foo.0", "bar")?;
+    run_test("[{foo: 'bar'}, {}]?.foo.1  | to nuon", "null")
 }
 
 // test that accessing a nonexistent row fails
 #[test]
 fn list_row_access_failure() -> TestResult {
     fail_test("[{foo: 'bar'}, {foo: 'baz'}].2", "")
+}
+
+#[test]
+fn list_row_optional_access_succeeds() -> TestResult {
+    run_test("[{foo: 'bar'}, {foo: 'baz'}]?.2  | to nuon", "null")
 }
 
 // regression test for an old bug
@@ -85,4 +126,94 @@ fn do_not_delve_too_deep_in_nested_lists() -> TestResult {
         "[[{foo: bar}]].foo",
         "did not find anything under this name",
     )
+}
+
+// tests for ListStreams, currently unused
+// TODO: re-enable these tests once we have a way to do streaming cell path access
+// I'm using a hack to create ListStreams for testing: pipe a list into `each`
+
+#[ignore = "cell path access doesn't handle ListStreams properly yet"]
+#[test]
+fn list_stream_single_field_success() -> TestResult {
+    run_test(
+        "[{foo: 'bar'} {foo: 'baz'}] | each {|i| $i } | get foo.0",
+        "bar",
+    )?;
+    run_test(
+        "[{foo: 'bar'} {foo: 'baz'}] | each {|i| $i } | get 0.foo",
+        "bar",
+    )?;
+    run_test(
+        "[{foo: 'bar'} {foo: 'baz'}] | each {|i| $i } | get 1.foo",
+        "baz",
+    )?;
+    run_test(
+        "[{foo: 'bar'} {foo: 'baz'}] | each {|i| $i } | get foo.1",
+        "baz",
+    )
+}
+
+#[ignore = "cell path access doesn't handle ListStreams properly yet"]
+#[test]
+fn list_stream_single_field_failure() -> TestResult {
+    fail_test(
+        "[{foo: 'bar'} {foo: 'baz'}] | each {|i| $i } | get asdf",
+        "",
+    )
+}
+
+#[ignore = "cell path access doesn't handle ListStreams properly yet"]
+#[test]
+fn list_stream_single_field_optional() -> TestResult {
+    run_test(
+        "[{foo: 'bar'} {foo: 'baz'}] | each {|i| $i } | get ?.asdf.0 | to nuon",
+        "null",
+    )?;
+    run_test(
+        "[{foo: 'bar'} {foo: 'baz'}] | each {|i| $i } | get 0?.asdf | to nuon",
+        "null",
+    )
+}
+
+#[ignore = "cell path access doesn't handle ListStreams properly yet"]
+#[test]
+fn jagged_liststream_access_fails() -> TestResult {
+    fail_test("[{foo: 'bar'} {}] | each {|i| $i } | get foo", "")?;
+    fail_test("[{} {foo: 'bar'}] | each {|i| $i } | get foo", "")
+}
+
+#[ignore = "cell path access doesn't handle ListStreams properly yet"]
+#[test]
+fn jagged_liststream_optional_access_succeeds() -> TestResult {
+    run_test("[{} {foo: 'bar'}] | each {|i| $i } | get ?.foo.1", "bar")?;
+    run_test("[{} {foo: 'bar'}] | each {|i| $i } | get 1?.foo", "bar")?;
+    run_test(
+        "[{} {foo: 'bar'}] | each {|i| $i } | get 0?.foo | to nuon",
+        "null",
+    )?;
+    run_test(
+        "[{} {foo: 'bar'}] | each {|i| $i } | get ?.foo.0 | to nuon",
+        "null",
+    )
+}
+
+// Tests for cell paths as used by `get`
+// Any cell path access (ex: `$foo.a`) can be rewritten with `get` (ex: `$foo | get a`)
+// This is mostly a straightforward translation, but it gets a little interesting because the first dot in a cell path is optional.
+// For example, `$foo?.a` can translate to `$foo | get ?.a` OR `$foo | get ?a`
+// `$foo.a` can translate to `$foo | get a` (the usual) or `$foo | get .a` (uncommon, but allowed)
+
+#[test]
+fn simple_get() -> TestResult {
+    run_test("{foo: 'bar'} | get foo", "bar")
+}
+
+#[test]
+fn get_with_prefixes() -> TestResult {
+    run_test("{foo: 'bar'} | get .foo", "bar")?;
+    run_test("{foo: 'bar'} | get ?foo", "bar")?;
+    run_test("{foo: 'bar'} | get ?.foo", "bar")?;
+
+    run_test("{} | get ?foo | to nuon", "null")?;
+    run_test("{} | get ?.foo | to nuon", "null")
 }
